% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata_merge.R
\name{stata_merge}
\alias{stata_merge}
\title{A merge Function}
\usage{
stata_merge(x, y)
}
\arguments{
\item{x}{data A}

\item{y}{data B}
}
\description{
equivalent of stata's "merge 1:1 ..." command to get the _merge indicator variable!
}
\examples{
stata_merge()

}
